refactor: collapse file barrel into file/index.ts
